Shooting Enemy Refactored
