+++ Assets/Scripts/Bullet.cs
@@ -19,11 +19,11 @@ public class Bullet : MonoBehaviour
 
         if (Enemy != null && owner != null && owner.tag == "Enemy")
         {
-            if (Enemy.GetComponent<Enemy>().isFacingRight)
+            if (Enemy.GetComponent<ShootingType>().isFacingRight)
             {
                 rb.velocity = transform.right * speed;
             }
-            else if (!Enemy.GetComponent<Enemy>().isFacingRight)
+            else if (!Enemy.GetComponent<ShootingType>().isFacingRight)
             {
                 rb.velocity = transform.right * speed * -1;
             }
