+++ Assets/Scripts/ShootingType.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 245d789ba3a790d469a9d7a51bf78a1f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
